implemented kafka consumer group
