+++ config-devel.toml
@@ -3,7 +3,6 @@ address = "kafka:29092"
 topic = "ccx.ocp.results"
 group = "aggregator"
 enabled = true
-save_offset = true
 enable_org_whitelist = false
 
 [server]
