+++ consumer/consts.go
@@ -1,0 +1,18 @@
+package consumer
+
+const (
+	// key for topic name used in structured log messages
+	topicKey = "topic"
+	// key for broker group name used in structured log messages
+	groupKey = "group"
+	// key for message offset used in structured log messages
+	offsetKey = "offset"
+	// key for message partition used in structured log messages
+	partitionKey = "partition"
+	// key for organization ID used in structured log messages
+	organizationKey = "organization"
+	// key for cluster ID used in structured log messages
+	clusterKey = "cluster"
+	// key for duration message type used in structured log messages
+	durationKey = "duration"
+)
