+++ tests/config1.toml
@@ -3,7 +3,6 @@ address = "localhost:29092"
 topic = "platform.results.ccx"
 group = "aggregator"
 enabled = false
-save_offset = false
 enable_org_whitelist = true
 
 [server]
